requirements.txt
certifi==2020.12.5
cffi==1.14.4
easydict==1.9
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pycparser==2.20
PyYAML==5.4.1
scipy==1.1.0
six==1.15.0
torch==0.4.1
torchvision==0.2.2
tqdm==4.19.9
